requirements.txt
# Web Framework
Flask==3.1.0
Werkzeug==3.1.3

# API Client
openai==1.68.2

# Audio Processing
numpy==1.26.4
soundfile==0.13.1

# Pre-install spacy and its dependencies to avoid build issues
spacy==3.7.2

# Kokoro TTS
kokoro==0.9.2

# Utilities
pathlib==1.0.1
python-dotenv==1.0.1